ecommerceapi/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.1
cryptography==42.0.6
Django==5.0.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-debug-toolbar==4.4.2
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
dr-scaffold==1.4.3
drf-yasg==1.21.7
idna==3.7
inflect==7.2.1
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
more-itertools==10.2.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
